requirements.txt
pyTelegramBotAPI==4.15.2
python-dotenv==1.0.0
APScheduler==3.10.4
Pillow==10.2.0